Separate out the test
